requirements.txt
altair==5.5.0
asttokens==2.2.1
attrs==25.1.0
Babel==2.12.1
backcall==0.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.1.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==7.2.1
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
narwhals==1.29.1
nest-asyncio==1.5.6
numpy==1.25.2
packaging==23.1
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.25.6
psutil==5.9.5
pure-eval==0.2.2
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.15.1
pyparsing==3.2.1
python-dateutil==2.8.2
pytz==2025.1
pywin32==306
pyzmq==25.1.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.2
stack-data==0.6.2
streamlit==1.43.1
tenacity==8.5.0
toml==0.10.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.6
zipp==3.21.0
